rename rider_devices_list subfolder
