+++ lib/widgets/pages/member_details/rider_devices_list/delete_device_button.dart
