+++ lib/widgets/pages/member_details/rider_devices_list/rider_devices_list_item.dart
